Fixing Network and cleaning up code

+++ Signal.cpp
@@ -21,7 +21,7 @@ void Signal::setClusterIndex(double clusterIndex) {
     m_clusterIndex = clusterIndex;
 }
 
-double Signal::getClusterIndex() {
+int Signal::getClusterIndex() {
     return m_clusterIndex;
 }
